+++ packages/workspace/src/browser/workspace-frontend-module.ts
@@ -44,6 +44,8 @@ import { WorkspaceDuplicateHandler } from './workspace-duplicate-handler';
 import { WorkspaceUtils } from './workspace-utils';
 import { WorkspaceCompareHandler } from './workspace-compare-handler';
 import { DiffService } from './diff-service';
+import { JsonSchemaContribution } from '@theia/core/lib/browser/json-schema-store';
+import { WorkspaceSchemaUpdater } from './workspace-schema-updater';
 
 export default new ContainerModule((bind: interfaces.Bind, unbind: interfaces.Unbind, isBound: interfaces.IsBound, rebind: interfaces.Rebind) => {
     bindWorkspacePreferences(bind);
@@ -91,4 +93,7 @@ export default new ContainerModule((bind: interfaces.Bind, unbind: interfaces.Un
     bind(QuickOpenWorkspace).toSelf().inSingletonScope();
 
     bind(WorkspaceUtils).toSelf().inSingletonScope();
+
+    bind(WorkspaceSchemaUpdater).toSelf().inSingletonScope();
+    bind(JsonSchemaContribution).toService(WorkspaceSchemaUpdater);
 });